+++ crates/runmat-vm/src/compiler/core.rs
@@ -1,8 +1,6 @@
-use crate::compiler::end_expr::{end_numeric_expr, expr_is_one, range_dynamic_end_spec};
 use crate::compiler::CompileError;
 use crate::functions::UserFunction;
-use crate::instr::{EmitLabel, EndExpr, Instr};
-use once_cell::sync::OnceCell;
+use crate::instr::{EmitLabel, Instr};
 use runmat_builtins::{self, Type};
 use runmat_hir::{HirExpr, HirExprKind, HirProgram, HirStmt};
 use std::collections::HashMap;
@@ -50,107 +48,8 @@ impl Drop for SpanGuard {
     }
 }
 
-struct StochasticEvolutionPlan<'a> {
-    state: runmat_hir::VarId,
-    drift: &'a HirExpr,
-    scale: &'a HirExpr,
-    steps: &'a HirExpr,
-}
-
-fn stochastic_evolution_disabled() -> bool {
-    static DISABLE: OnceCell<bool> = OnceCell::new();
-    *DISABLE.get_or_init(|| {
-        std::env::var("RUNMAT_DISABLE_STOCHASTIC_EVOLUTION")
-            .map(|v| matches!(v.trim().to_ascii_lowercase().as_str(), "1" | "true" | "yes"))
-            .unwrap_or(false)
-    })
-}
-
-fn is_randn_call(expr: &HirExpr) -> bool {
-    match &expr.kind {
-        runmat_hir::HirExprKind::FuncCall(name, _) => name.eq_ignore_ascii_case("randn"),
-        _ => false,
-    }
-}
-
-fn expr_supports_inline_emit(expr: &HirExpr) -> bool {
-    match &expr.kind {
-        HirExprKind::FuncCall(name, _) => !runmat_builtins::suppresses_auto_output(name),
-        _ => true,
-    }
-}
-
-fn matches_var(expr: &HirExpr, var: runmat_hir::VarId) -> bool {
-    matches!(expr.kind, runmat_hir::HirExprKind::Var(id) if id == var)
-}
-
-fn extract_drift_and_scale(
-    expr: &HirExpr,
-    state_var: runmat_hir::VarId,
-    z_var: runmat_hir::VarId,
-) -> Option<(&HirExpr, &HirExpr)> {
-    use runmat_hir::HirExprKind as EK;
-    use runmat_parser::BinOp;
-
-    let (maybe_state_side, maybe_exp_side) = match &expr.kind {
-        EK::Binary(lhs, BinOp::ElemMul, rhs) => (lhs.as_ref(), rhs.as_ref()),
-        _ => return None,
-    };
-
-    let exp_side = if matches_var(maybe_state_side, state_var) && is_exp_call(maybe_exp_side) {
-        maybe_exp_side
-    } else if matches_var(maybe_exp_side, state_var) && is_exp_call(maybe_state_side) {
-        maybe_state_side
-    } else {
-        return None;
-    };
-
-    let exp_arg = match &exp_side.kind {
-        EK::FuncCall(name, args) if name.eq_ignore_ascii_case("exp") && args.len() == 1 => &args[0],
-        _ => return None,
-    };
-
-    match &exp_arg.kind {
-        EK::Binary(lhs, BinOp::Add, rhs) => {
-            if let Some(scale_expr) = extract_scale_term(lhs, z_var) {
-                Some((rhs.as_ref(), scale_expr))
-            } else if let Some(scale_expr) = extract_scale_term(rhs, z_var) {
-                Some((lhs.as_ref(), scale_expr))
-            } else {
-                None
-            }
-        }
-        _ => None,
-    }
-}
-
-fn extract_scale_term(expr: &HirExpr, z_var: runmat_hir::VarId) -> Option<&HirExpr> {
-    use runmat_hir::HirExprKind as EK;
-    use runmat_parser::BinOp;
-
-    match &expr.kind {
-        EK::Binary(lhs, BinOp::ElemMul, rhs) => {
-            if matches_var(lhs, z_var) {
-                Some(rhs.as_ref())
-            } else if matches_var(rhs, z_var) {
-                Some(lhs.as_ref())
-            } else {
-                None
-            }
-        }
-        _ => None,
-    }
-}
-
-fn is_exp_call(expr: &HirExpr) -> bool {
-    matches!(
-        &expr.kind,
-        runmat_hir::HirExprKind::FuncCall(name, _) if name.eq_ignore_ascii_case("exp")
-    )
-}
-
 impl Compiler {
-    fn normalize_class_literal_name(raw: &str) -> String {
+    pub(crate) fn normalize_class_literal_name(raw: &str) -> String {
         if raw.len() >= 2 {
             let bytes = raw.as_bytes();
             let first = bytes[0] as char;
@@ -162,20 +61,7 @@ impl Compiler {
         raw.to_string()
     }
 
-    fn compile_stochastic_evolution(
-        &mut self,
-        plan: StochasticEvolutionPlan<'_>,
-    ) -> Result<(), CompileError> {
-        self.emit(Instr::LoadVar(plan.state.0));
-        self.compile_expr(plan.drift)?;
-        self.compile_expr(plan.scale)?;
-        self.compile_expr(plan.steps)?;
-        self.emit(Instr::StochasticEvolution);
-        self.emit(Instr::StoreVar(plan.state.0));
-        Ok(())
-    }
-
-    fn emit_multiassign_outputs(&mut self, vars: &[Option<runmat_hir::VarId>]) {
+    pub(crate) fn emit_multiassign_outputs(&mut self, vars: &[Option<runmat_hir::VarId>]) {
         for v in vars.iter().flatten() {
             self.emit(Instr::EmitVar {
                 var_index: v.0,
@@ -184,116 +70,6 @@ impl Compiler {
         }
     }
 
-    fn detect_stochastic_evolution<'a>(
-        &self,
-        expr: &'a HirExpr,
-        body: &'a [HirStmt],
-    ) -> Option<StochasticEvolutionPlan<'a>> {
-        if stochastic_evolution_disabled() {
-            return None;
-        }
-        use runmat_hir::HirExprKind as EK;
-        match &expr.kind {
-            EK::Range(start, step, end) => {
-                if !expr_is_one(start) {
-                    return None;
-                }
-                if let Some(step_expr) = step {
-                    if !expr_is_one(step_expr) {
-                        return None;
-                    }
-                }
-                if body.len() != 2 {
-                    return None;
-                }
-                let (z_var, randn_expr) = match &body[0] {
-                    HirStmt::Assign(var, expr, _, _) => (*var, expr),
-                    _ => return None,
-                };
-                if !is_randn_call(randn_expr) {
-                    return None;
-                }
-                let (state_var, update_expr) = match &body[1] {
-                    HirStmt::Assign(var, expr, _, _) => (*var, expr),
-                    _ => return None,
-                };
-                let (drift, scale) = extract_drift_and_scale(update_expr, state_var, z_var)?;
-                Some(StochasticEvolutionPlan {
-                    state: state_var,
-                    drift,
-                    scale,
-                    steps: end,
-                })
-            }
-            _ => None,
-        }
-    }
-    fn attr_access_from_str(s: &str) -> runmat_builtins::Access {
-        match s.to_ascii_lowercase().as_str() {
-            "private" => runmat_builtins::Access::Private,
-            _ => runmat_builtins::Access::Public,
-        }
-    }
-    fn parse_prop_attrs(attrs: &Vec<runmat_parser::Attr>) -> (bool, bool, String, String) {
-        // Defaults
-        let mut is_static = false;
-        let mut is_dependent = false;
-        let mut get_acc = runmat_builtins::Access::Public;
-        let mut set_acc = runmat_builtins::Access::Public;
-        for a in attrs {
-            if a.name.eq_ignore_ascii_case("Static") {
-                is_static = true;
-            }
-            if a.name.eq_ignore_ascii_case("Dependent") {
-                is_dependent = true;
-            }
-            if a.name.eq_ignore_ascii_case("Access") {
-                if let Some(v) = &a.value {
-                    let acc = Self::attr_access_from_str(v.trim_matches('\'').trim());
-                    get_acc = acc.clone();
-                    set_acc = acc;
-                }
-            }
-            if a.name.eq_ignore_ascii_case("GetAccess") {
-                if let Some(v) = &a.value {
-                    get_acc = Self::attr_access_from_str(v.trim_matches('\'').trim());
-                }
-            }
-            if a.name.eq_ignore_ascii_case("SetAccess") {
-                if let Some(v) = &a.value {
-                    set_acc = Self::attr_access_from_str(v.trim_matches('\'').trim());
-                }
-            }
-        }
-        let gs = match get_acc {
-            runmat_builtins::Access::Private => "private".to_string(),
-            _ => "public".to_string(),
-        };
-        let ss = match set_acc {
-            runmat_builtins::Access::Private => "private".to_string(),
-            _ => "public".to_string(),
-        };
-        (is_static, is_dependent, gs, ss)
-    }
-    fn parse_method_attrs(attrs: &Vec<runmat_parser::Attr>) -> (bool, String) {
-        let mut is_static = false;
-        let mut access = runmat_builtins::Access::Public;
-        for a in attrs {
-            if a.name.eq_ignore_ascii_case("Static") {
-                is_static = true;
-            }
-            if a.name.eq_ignore_ascii_case("Access") {
-                if let Some(v) = &a.value {
-                    access = Self::attr_access_from_str(v.trim_matches('\'').trim());
-                }
-            }
-        }
-        let acc_str = match access {
-            runmat_builtins::Access::Private => "private".to_string(),
-            _ => "public".to_string(),
-        };
-        (is_static, acc_str)
-    }
     pub fn new(prog: &HirProgram) -> Self {
         let mut max_var = 0;
         fn visit_expr(expr: &HirExpr, max: &mut usize) {
@@ -500,7 +276,11 @@ impl Compiler {
         pc
     }
 
-    fn emit_call_with_arg_spans(&mut self, instr: Instr, arg_spans: &[runmat_hir::Span]) -> usize {
+    pub(crate) fn emit_call_with_arg_spans(
+        &mut self,
+        instr: Instr,
+        arg_spans: &[runmat_hir::Span],
+    ) -> usize {
         let pc = self.emit(instr);
         if !arg_spans.is_empty() {
             if let Some(slot) = self.call_arg_spans.get_mut(pc) {
@@ -561,1415 +341,11 @@ impl Compiler {
 
     pub fn compile_stmt(&mut self, stmt: &HirStmt) -> Result<(), CompileError> {
         let _span_guard = SpanGuard::new(self, stmt.span());
-        match stmt {
-            HirStmt::ExprStmt(expr, suppressed, _) => {
-                self.compile_expr(expr)?;
-                if !suppressed && expr_supports_inline_emit(expr) {
-                    let label = label_for_expr(expr);
-                    self.emit(Instr::EmitStackTop { label });
-                }
-                self.emit(Instr::Pop);
-            }
-            HirStmt::Assign(id, expr, suppressed, _) => {
-                self.compile_expr(expr)?;
-                self.emit(Instr::StoreVar(id.0));
-                if !suppressed {
-                    self.emit(Instr::EmitVar {
-                        var_index: id.0,
-                        label: EmitLabel::Var(id.0),
-                    });
-                }
-            }
-            HirStmt::If {
-                cond,
-                then_body,
-                elseif_blocks,
-                else_body,
-                ..
-            } => {
-                self.compile_expr(cond)?;
-                let mut last_jump = self.emit(Instr::JumpIfFalse(usize::MAX));
-                for s in then_body {
-                    self.compile_stmt(s)?;
-                }
-                let mut end_jumps = Vec::new();
-                end_jumps.push(self.emit(Instr::Jump(usize::MAX)));
-                for (c, b) in elseif_blocks {
-                    self.patch(last_jump, Instr::JumpIfFalse(self.instructions.len()));
-                    self.compile_expr(c)?;
-                    last_jump = self.emit(Instr::JumpIfFalse(usize::MAX));
-                    for s in b {
-                        self.compile_stmt(s)?;
-                    }
-                    end_jumps.push(self.emit(Instr::Jump(usize::MAX)));
-                }
-                self.patch(last_jump, Instr::JumpIfFalse(self.instructions.len()));
-                if let Some(body) = else_body {
-                    for s in body {
-                        self.compile_stmt(s)?;
-                    }
-                }
-                let end = self.instructions.len();
-                for j in end_jumps {
-                    self.patch(j, Instr::Jump(end));
-                }
-            }
-            HirStmt::While { cond, body, .. } => {
-                let start = self.instructions.len();
-                self.compile_expr(cond)?;
-                let jump_end = self.emit(Instr::JumpIfFalse(usize::MAX));
-                let labels = LoopLabels {
-                    break_jumps: Vec::new(),
-                    continue_jumps: Vec::new(),
-                };
-                self.loop_stack.push(labels);
-                for s in body {
-                    self.compile_stmt(s)?;
-                }
-                let labels = self.loop_stack.pop().unwrap();
-                for j in labels.continue_jumps {
-                    self.patch(j, Instr::Jump(start));
-                }
-                self.emit(Instr::Jump(start));
-                let end = self.instructions.len();
-                self.patch(jump_end, Instr::JumpIfFalse(end));
-                for j in labels.break_jumps {
-                    self.patch(j, Instr::Jump(end));
-                }
-            }
-            HirStmt::For {
-                var, expr, body, ..
-            } => {
-                if let Some(plan) = self.detect_stochastic_evolution(expr, body) {
-                    self.compile_stochastic_evolution(plan)?;
-                    return Ok(());
-                }
-                if let HirExprKind::Range(start, step, end) = &expr.kind {
-                    // Initialize loop variable, end, and step
-                    self.compile_expr(start)?;
-                    self.emit(Instr::StoreVar(var.0));
-                    self.compile_expr(end)?;
-                    let end_var = self.alloc_temp();
-                    self.emit(Instr::StoreVar(end_var));
-                    let step_var = self.alloc_temp();
-                    if let Some(step_expr) = step {
-                        self.compile_expr(step_expr)?;
-                        self.emit(Instr::StoreVar(step_var));
-                    } else {
-                        self.emit(Instr::LoadConst(1.0));
-                        self.emit(Instr::StoreVar(step_var));
-                    }
-
-                    let loop_start = self.instructions.len();
-
-                    // If step == 0 -> terminate loop immediately
-                    self.emit(Instr::LoadVar(step_var));
-                    self.emit(Instr::LoadConst(0.0));
-                    self.emit(Instr::Equal);
-                    let j_step_zero_skip = self.emit(Instr::JumpIfFalse(usize::MAX));
-                    let jump_end = self.emit(Instr::Jump(usize::MAX));
-                    let after_step_zero_check = self.instructions.len();
-                    self.patch(j_step_zero_skip, Instr::JumpIfFalse(after_step_zero_check));
-
-                    // Determine condition based on sign(step)
-                    // if step >= 0: cond = var <= end else cond = var >= end
-                    self.emit(Instr::LoadVar(step_var));
-                    self.emit(Instr::LoadConst(0.0));
-                    self.emit(Instr::GreaterEqual);
-                    let j_neg_branch = self.emit(Instr::JumpIfFalse(usize::MAX));
-                    // positive step: var <= end
-                    self.emit(Instr::LoadVar(var.0));
-                    self.emit(Instr::LoadVar(end_var));
-                    self.emit(Instr::LessEqual);
-                    let j_exit_pos = self.emit(Instr::JumpIfFalse(usize::MAX));
-                    let j_cond_done = self.emit(Instr::Jump(usize::MAX));
-                    let neg_branch = self.instructions.len();
-                    self.patch(j_neg_branch, Instr::JumpIfFalse(neg_branch));
-                    // negative step: var >= end
-                    self.emit(Instr::LoadVar(var.0));
-                    self.emit(Instr::LoadVar(end_var));
-                    self.emit(Instr::GreaterEqual);
-                    let j_exit_neg = self.emit(Instr::JumpIfFalse(usize::MAX));
-                    let cond_done = self.instructions.len();
-                    self.patch(j_cond_done, Instr::Jump(cond_done));
-
-                    // At this point, loop condition is satisfied; execute body
-                    self.loop_stack.push(LoopLabels {
-                        break_jumps: Vec::new(),
-                        continue_jumps: Vec::new(),
-                    });
-                    for s in body {
-                        self.compile_stmt(s)?;
-                    }
-                    let labels = self.loop_stack.pop().unwrap();
-                    for j in labels.continue_jumps {
-                        self.patch(j, Instr::Jump(self.instructions.len()));
-                    }
-                    // increment: var = var + step
-                    self.emit(Instr::LoadVar(var.0));
-                    self.emit(Instr::LoadVar(step_var));
-                    self.emit(Instr::Add);
-                    self.emit(Instr::StoreVar(var.0));
-                    self.emit(Instr::Jump(loop_start));
-
-                    // Exit jump targets
-                    let end_pc = self.instructions.len();
-                    self.patch(jump_end, Instr::Jump(end_pc));
-                    self.patch(j_exit_pos, Instr::JumpIfFalse(end_pc));
-                    self.patch(j_exit_neg, Instr::JumpIfFalse(end_pc));
-                    for j in labels.break_jumps {
-                        self.patch(j, Instr::Jump(end_pc));
-                    }
-                } else {
-                    return Err(self.compile_error("for loop expects range"));
-                }
-            }
-            HirStmt::Break(_) => {
-                if self.loop_stack.is_empty() {
-                    return Err(self.compile_error("break outside loop"));
-                }
-                let idx = self.emit(Instr::Jump(usize::MAX));
-                if let Some(labels) = self.loop_stack.last_mut() {
-                    labels.break_jumps.push(idx);
-                }
-            }
-            HirStmt::Continue(_) => {
-                if self.loop_stack.is_empty() {
-                    return Err(self.compile_error("continue outside loop"));
-                }
-                let idx = self.emit(Instr::Jump(usize::MAX));
-                if let Some(labels) = self.loop_stack.last_mut() {
-                    labels.continue_jumps.push(idx);
-                }
-            }
-            HirStmt::Return(_) => {
-                self.emit(Instr::Return);
-            }
-            HirStmt::Function {
-                name,
-                params,
-                outputs,
-                body,
-                has_varargin,
-                has_varargout,
-                ..
-            } => self.compile_function_stmt(
-                name,
-                params,
-                outputs,
-                body,
-                *has_varargin,
-                *has_varargout,
-            )?,
-            HirStmt::Switch {
-                expr,
-                cases,
-                otherwise,
-                ..
-            } => {
-                let temp_id = self.alloc_temp();
-                self.compile_expr(expr)?;
-                self.emit(Instr::StoreVar(temp_id));
-                let mut end_jumps: Vec<usize> = Vec::new();
-                let mut next_case_jump_to_here: Option<usize> = None;
-                for (case_expr, body) in cases {
-                    if let Some(j) = next_case_jump_to_here.take() {
-                        self.patch(j, Instr::JumpIfFalse(self.instructions.len()));
-                    }
-                    self.emit(Instr::LoadVar(temp_id));
-                    self.compile_expr(case_expr)?;
-                    self.emit(Instr::Equal);
-                    let jmp = self.emit(Instr::JumpIfFalse(usize::MAX));
-                    for s in body {
-                        self.compile_stmt(s)?;
-                    }
-                    end_jumps.push(self.emit(Instr::Jump(usize::MAX)));
-                    next_case_jump_to_here = Some(jmp);
-                }
-                let otherwise_start = self.instructions.len();
-                if let Some(j) = next_case_jump_to_here.take() {
-                    self.patch(j, Instr::JumpIfFalse(otherwise_start));
-                }
-                if let Some(body) = otherwise {
-                    for s in body {
-                        self.compile_stmt(s)?;
-                    }
-                }
-                let end = self.instructions.len();
-                for j in end_jumps {
-                    self.patch(j, Instr::Jump(end));
-                }
-            }
-            HirStmt::TryCatch {
-                try_body,
-                catch_var,
-                catch_body,
-                ..
-            } => {
-                // Reserve slot for EnterTry with placeholder
-                let enter_idx = self.emit(Instr::EnterTry(usize::MAX, catch_var.map(|v| v.0)));
-                // Compile try body
-                for s in try_body {
-                    self.compile_stmt(s)?;
-                }
-                // On normal completion, pop try frame and jump past catch
-                self.emit(Instr::PopTry);
-                let jmp_end = self.emit(Instr::Jump(usize::MAX));
-                // Catch block starts here
-                let catch_pc = self.instructions.len();
-                // Patch EnterTry with catch_pc
-                self.patch(enter_idx, Instr::EnterTry(catch_pc, catch_var.map(|v| v.0)));
-                // Compile catch body
-                for s in catch_body {
-                    self.compile_stmt(s)?;
-                }
-                let end_pc = self.instructions.len();
-                self.patch(jmp_end, Instr::Jump(end_pc));
-            }
-            HirStmt::AssignLValue(lv, rhs, _, _) => self.compile_assign_lvalue(lv, rhs)?,
-            HirStmt::Global(vars, _) => {
-                let ids: Vec<usize> = vars.iter().map(|(v, _n)| v.0).collect();
-                let names: Vec<String> = vars.iter().map(|(_v, n)| n.clone()).collect();
-                self.emit(Instr::DeclareGlobalNamed(ids, names));
-            }
-            HirStmt::Persistent(vars, _) => {
-                let ids: Vec<usize> = vars.iter().map(|(v, _n)| v.0).collect();
-                let names: Vec<String> = vars.iter().map(|(_v, n)| n.clone()).collect();
-                self.emit(Instr::DeclarePersistentNamed(ids, names));
-            }
-            HirStmt::Import { path, wildcard, .. } => {
-                self.emit(Instr::RegisterImport {
-                    path: path.clone(),
-                    wildcard: *wildcard,
-                });
-            }
-            HirStmt::ClassDef {
-                name,
-                super_class,
-                members,
-                ..
-            } => {
-                // Synthesize a minimal RegisterClass instruction by extracting property names and method names
-                let mut props: Vec<(String, bool, String, String)> = Vec::new();
-                let mut methods: Vec<(String, String, bool, String)> = Vec::new();
-                for m in members {
-                    match m {
-                        runmat_hir::HirClassMember::Properties { names, attributes } => {
-                            let (is_static, is_dependent, get_access, set_access) =
-                                Self::parse_prop_attrs(attributes);
-                            // Encode dependent flag by prefixing name with "@dep:"; VM will strip and set flag.
-                            for n in names {
-                                let enc = if is_dependent {
-                                    format!("@dep:{n}")
-                                } else {
-                                    n.clone()
-                                };
-                                props.push((
-                                    enc,
-                                    is_static,
-                                    get_access.clone(),
-                                    set_access.clone(),
-                                ));
-                            }
-                        }
-                        runmat_hir::HirClassMember::Methods { body, attributes } => {
-                            let (is_static, access) = Self::parse_method_attrs(attributes);
-                            for s in body {
-                                if let runmat_hir::HirStmt::Function { name: mname, .. } = s {
-                                    methods.push((
-                                        mname.clone(),
-                                        mname.clone(),
-                                        is_static,
-                                        access.clone(),
-                                    ));
-                                }
-                            }
-                        }
-                        _ => {}
-                    }
-                }
-                self.emit(Instr::RegisterClass {
-                    name: name.clone(),
-                    super_class: super_class.clone(),
-                    properties: props,
-                    methods,
-                });
-            }
-            HirStmt::MultiAssign(vars, expr, suppressed, _) => {
-                // Compile RHS once; if function call or value, arrange to extract multiple
-                match &expr.kind {
-                    HirExprKind::FuncCall(name, args) => {
-                        let call_arg_spans: Vec<runmat_hir::Span> =
-                            args.iter().map(|a| a.span).collect();
-                        if self.functions.contains_key(name) {
-                            for arg in args {
-                                self.compile_expr(arg)?;
-                            }
-                            // Emit multi-call to request N outputs
-                            self.emit_call_with_arg_spans(
-                                Instr::CallFunctionMulti(name.clone(), args.len(), vars.len()),
-                                &call_arg_spans,
-                            );
-                            self.emit(Instr::Unpack(vars.len()));
-                            // Store outputs in order
-                            for (_i, var) in vars.iter().enumerate().rev() {
-                                if let Some(v) = var {
-                                    self.emit(Instr::StoreVar(v.0));
-                                } else {
-                                    self.emit(Instr::Pop);
-                                }
-                            }
-                        } else {
-                            // Builtin or unknown: treat as single return value
-                            for arg in args {
-                                self.compile_expr(arg)?;
-                            }
-                            self.emit_call_with_arg_spans(
-                                Instr::CallBuiltin(name.clone(), args.len()),
-                                &call_arg_spans,
-                            );
-                            self.emit(Instr::Unpack(vars.len()));
-                            for (_i, var) in vars.iter().enumerate().rev() {
-                                if let Some(v) = var {
-                                    self.emit(Instr::StoreVar(v.0));
-                                } else {
-                                    self.emit(Instr::Pop);
-                                }
-                            }
-                        }
-                    }
-                    HirExprKind::IndexCell(base, indices) => {
-                        // Support comma-list expansion from cell indexing: [a,b,...] = C{idx}
-                        self.compile_expr(base)?;
-                        for index in indices {
-                            self.compile_expr(index)?;
-                        }
-                        // Expand into N outputs
-                        self.emit(Instr::IndexCellExpand(indices.len(), vars.len()));
-                        for (_i, var) in vars.iter().enumerate().rev() {
-                            if let Some(v) = var {
-                                self.emit(Instr::StoreVar(v.0));
-                            } else {
-                                self.emit(Instr::Pop);
-                            }
-                        }
-                    }
-                    _ => {
-                        // Non-call: assign expr to first non-placeholder, zeros to remaining non-placeholders
-                        let first_real = vars.iter().position(|v| v.is_some());
-                        if let Some(first_idx) = first_real {
-                            self.compile_expr(expr)?;
-                            if let Some(Some(first_var)) = vars.get(first_idx) {
-                                self.emit(Instr::StoreVar(first_var.0));
-                            }
-                        }
-                        for (i, var) in vars.iter().enumerate() {
-                            if Some(i) == first_real {
-                                continue;
-                            }
-                            if let Some(v) = var {
-                                self.emit(Instr::LoadConst(0.0));
-                                self.emit(Instr::StoreVar(v.0));
-                            }
-                        }
-                    }
-                }
-                if !suppressed {
-                    self.emit_multiassign_outputs(vars);
-                }
-            }
-        }
-        Ok(())
+        self.compile_stmt_impl(stmt)
     }
 
     pub fn compile_expr(&mut self, expr: &HirExpr) -> Result<(), CompileError> {
         let _span_guard = SpanGuard::new(self, expr.span);
-        match &expr.kind {
-            HirExprKind::Number(n) => {
-                let val: f64 = n.parse().map_err(|_| "invalid number")?;
-                self.emit(Instr::LoadConst(val));
-            }
-            HirExprKind::String(s) => {
-                if s.starts_with('"') && s.ends_with('"') {
-                    // String scalar
-                    let inner = &s[1..s.len() - 1];
-                    let clean = inner.replace("\"\"", "\"");
-                    self.emit(Instr::LoadString(clean));
-                } else if s.starts_with('\'') && s.ends_with('\'') {
-                    // Char vector -> CharArray row
-                    let inner = &s[1..s.len() - 1];
-                    let clean = inner.replace("''", "'");
-                    // Encode as CharArray(1, len)
-                    self.emit(Instr::LoadCharRow(clean));
-                } else {
-                    self.emit(Instr::LoadString(s.clone()));
-                }
-            }
-            HirExprKind::Var(id) => {
-                self.emit(Instr::LoadVar(id.0));
-            }
-            // Fallback path for unqualified static property via Class.* imports: treat bare identifier as Class.prop
-            // Only when not a known var/const/func; handled earlier in HIR, so we also handle here for robustness via Member lowering
-            // Note: HIR would have errored on undefined variable; we resolve at compile time before that for function calls,
-            // but for expressions like `v = staticValue;` under `import Point.*`, we handle here by probing imports.
-            HirExprKind::Constant(name) => {
-                let constants = runmat_builtins::constants();
-                if let Some(constant) = constants.iter().find(|c| c.name == name) {
-                    match &constant.value {
-                        runmat_builtins::Value::Num(val) => {
-                            self.emit(Instr::LoadConst(*val));
-                        }
-                        runmat_builtins::Value::Complex(re, im) => {
-                            self.emit(Instr::LoadComplex(*re, *im));
-                        }
-                        runmat_builtins::Value::Bool(b) => {
-                            self.emit(Instr::LoadBool(*b));
-                        }
-                        _ => {
-                            return Err(self.compile_error(format!(
-                                "Constant {name} is not a number or boolean"
-                            )));
-                        }
-                    }
-                } else {
-                    // Try resolving as unqualified static property via Class.* imports (multi-segment)
-                    let mut classes: Vec<String> = Vec::new();
-                    for (path, wildcard) in &self.imports {
-                        if !*wildcard {
-                            continue;
-                        }
-                        if path.is_empty() {
-                            continue;
-                        }
-                        let mut cls = String::new();
-                        for (i, part) in path.iter().enumerate() {
-                            if i > 0 {
-                                cls.push('.');
-                            }
-                            cls.push_str(part);
-                        }
-                        if let Some((p, _owner)) = runmat_builtins::lookup_property(&cls, name) {
-                            if p.is_static {
-                                classes.push(cls.clone());
-                            }
-                        }
-                    }
-                    if classes.len() > 1 {
-                        return Err(self.compile_error(format!(
-                            "ambiguous unqualified static property '{}' via Class.* imports: {}",
-                            name,
-                            classes.join(", ")
-                        )));
-                    }
-                    if classes.len() == 1 {
-                        self.emit(Instr::LoadStaticProperty(classes.remove(0), name.clone()));
-                        return Ok(());
-                    }
-                    return Err(
-                        self.compile_error(format!("Unknown constant or static property: {name}"))
-                    );
-                }
-            }
-            HirExprKind::Unary(op, e) => {
-                self.compile_expr(e)?;
-                match op {
-                    runmat_parser::UnOp::Plus => {
-                        self.emit(Instr::UPlus);
-                    }
-                    runmat_parser::UnOp::Minus => {
-                        self.emit(Instr::Neg);
-                    }
-                    runmat_parser::UnOp::Transpose => {
-                        self.emit(Instr::ConjugateTranspose);
-                    }
-                    runmat_parser::UnOp::NonConjugateTranspose => {
-                        self.emit(Instr::Transpose);
-                    }
-                    runmat_parser::UnOp::Not => {
-                        self.emit(Instr::CallBuiltin("not".to_string(), 1));
-                    }
-                }
-            }
-            HirExprKind::Binary(a, op, b) => {
-                use runmat_parser::BinOp;
-                match op {
-                    BinOp::AndAnd => {
-                        self.compile_expr(a)?;
-                        self.emit(Instr::LoadConst(0.0));
-                        self.emit(Instr::NotEqual);
-                        let j_false = self.emit(Instr::JumpIfFalse(usize::MAX));
-                        self.compile_expr(b)?;
-                        self.emit(Instr::LoadConst(0.0));
-                        self.emit(Instr::NotEqual);
-                        let end = self.emit(Instr::Jump(usize::MAX));
-                        let after_cond = self.instructions.len();
-                        self.patch(j_false, Instr::JumpIfFalse(after_cond));
-                        self.emit(Instr::LoadConst(0.0));
-                        let end_pc = self.instructions.len();
-                        self.patch(end, Instr::Jump(end_pc));
-                    }
-                    BinOp::OrOr => {
-                        self.compile_expr(a)?;
-                        self.emit(Instr::LoadConst(0.0));
-                        self.emit(Instr::NotEqual);
-                        let j_true = self.emit(Instr::JumpIfFalse(usize::MAX));
-                        self.emit(Instr::LoadConst(1.0));
-                        let end = self.emit(Instr::Jump(usize::MAX));
-                        let after_check = self.instructions.len();
-                        self.patch(j_true, Instr::JumpIfFalse(after_check));
-                        self.compile_expr(b)?;
-                        self.emit(Instr::LoadConst(0.0));
-                        self.emit(Instr::NotEqual);
-                        let end_pc = self.instructions.len();
-                        self.patch(end, Instr::Jump(end_pc));
-                    }
-                    BinOp::BitAnd => {
-                        self.compile_expr(a)?;
-                        self.emit(Instr::LoadConst(0.0));
-                        self.emit(Instr::CallBuiltin("ne".to_string(), 2));
-                        self.compile_expr(b)?;
-                        self.emit(Instr::LoadConst(0.0));
-                        self.emit(Instr::CallBuiltin("ne".to_string(), 2));
-                        self.emit(Instr::ElemMul);
-                    }
-                    BinOp::BitOr => {
-                        self.compile_expr(a)?;
-                        self.emit(Instr::LoadConst(0.0));
-                        self.emit(Instr::CallBuiltin("ne".to_string(), 2));
-                        self.compile_expr(b)?;
-                        self.emit(Instr::LoadConst(0.0));
-                        self.emit(Instr::CallBuiltin("ne".to_string(), 2));
-                        self.emit(Instr::Add);
-                        self.emit(Instr::LoadConst(0.0));
-                        self.emit(Instr::CallBuiltin("ne".to_string(), 2));
-                        // Stay element-wise by folding through elementwise ops
-                    }
-                    _ => {
-                        self.compile_expr(a)?;
-                        self.compile_expr(b)?;
-                        match op {
-                            BinOp::Add => {
-                                self.emit(Instr::Add);
-                            }
-                            BinOp::Sub => {
-                                self.emit(Instr::Sub);
-                            }
-                            BinOp::Mul => {
-                                self.emit(Instr::Mul);
-                            }
-                            BinOp::RightDiv => {
-                                self.emit(Instr::RightDiv);
-                            }
-                            BinOp::LeftDiv => {
-                                self.emit(Instr::LeftDiv);
-                            }
-                            BinOp::Pow => {
-                                self.emit(Instr::Pow);
-                            }
-                            BinOp::ElemMul => {
-                                self.emit(Instr::ElemMul);
-                            }
-                            BinOp::ElemDiv => {
-                                self.emit(Instr::ElemDiv);
-                            }
-                            BinOp::ElemLeftDiv => {
-                                self.emit(Instr::ElemLeftDiv);
-                            }
-                            BinOp::ElemPow => {
-                                self.emit(Instr::ElemPow);
-                            }
-                            BinOp::Equal => {
-                                self.emit(Instr::Equal);
-                            }
-                            BinOp::NotEqual => {
-                                self.emit(Instr::NotEqual);
-                            }
-                            BinOp::Less => {
-                                self.emit(Instr::Less);
-                            }
-                            BinOp::LessEqual => {
-                                self.emit(Instr::LessEqual);
-                            }
-                            BinOp::Greater => {
-                                self.emit(Instr::Greater);
-                            }
-                            BinOp::GreaterEqual => {
-                                self.emit(Instr::GreaterEqual);
-                            }
-                            BinOp::Colon => {
-                                return Err(self.compile_error("colon operator not supported"));
-                            }
-                            _ => unreachable!(),
-                        }
-                    }
-                }
-            }
-            HirExprKind::Range(start, step, end) => {
-                self.compile_expr(start)?;
-                if let Some(step) = step {
-                    self.compile_expr(step)?;
-                    self.compile_expr(end)?;
-                    self.emit(Instr::CreateRange(true));
-                } else {
-                    self.compile_expr(end)?;
-                    self.emit(Instr::CreateRange(false));
-                }
-            }
-            HirExprKind::FuncCall(name, args) => {
-                let call_arg_spans: Vec<runmat_hir::Span> = args.iter().map(|a| a.span).collect();
-                // Special-case: feval(f, a1, a2, ...) compiles to VM feval to access user functions/closures
-                if name == "feval" {
-                    if args.is_empty() {
-                        return Err(self.compile_error("feval: missing function argument"));
-                    }
-                    // Push function value first
-                    self.compile_expr(&args[0])?;
-                    let rest = &args[1..];
-                    let has_expand = rest
-                        .iter()
-                        .any(|a| matches!(a.kind, HirExprKind::IndexCell(_, _)));
-                    if has_expand {
-                        let mut specs: Vec<crate::instr::ArgSpec> = Vec::with_capacity(rest.len());
-                        for arg in rest {
-                            if let HirExprKind::IndexCell(base, indices) = &arg.kind {
-                                let is_expand_all = indices.len() == 1
-                                    && matches!(indices[0].kind, HirExprKind::Colon);
-                                if is_expand_all {
-                                    specs.push(crate::instr::ArgSpec {
-                                        is_expand: true,
-                                        num_indices: 0,
-                                        expand_all: true,
-                                    });
-                                    self.compile_expr(base)?;
-                                } else {
-                                    specs.push(crate::instr::ArgSpec {
-                                        is_expand: true,
-                                        num_indices: indices.len(),
-                                        expand_all: false,
-                                    });
-                                    self.compile_expr(base)?;
-                                    for i in indices {
-                                        self.compile_expr(i)?;
-                                    }
-                                }
-                            } else {
-                                specs.push(crate::instr::ArgSpec {
-                                    is_expand: false,
-                                    num_indices: 0,
-                                    expand_all: false,
-                                });
-                                self.compile_expr(arg)?;
-                            }
-                        }
-                        self.emit_call_with_arg_spans(
-                            Instr::CallFevalExpandMulti(specs),
-                            &call_arg_spans,
-                        );
-                    } else {
-                        for arg in rest {
-                            self.compile_expr(arg)?;
-                        }
-                        self.emit_call_with_arg_spans(
-                            Instr::CallFeval(rest.len()),
-                            &call_arg_spans,
-                        );
-                    }
-                    return Ok(());
-                }
-                let has_any_expand = args
-                    .iter()
-                    .any(|a| matches!(a.kind, HirExprKind::IndexCell(_, _)));
-                if self.functions.contains_key(name) {
-                    // existing path
-                    if has_any_expand {
-                        let mut specs: Vec<crate::instr::ArgSpec> = Vec::with_capacity(args.len());
-                        for arg in args {
-                            if let HirExprKind::IndexCell(base, indices) = &arg.kind {
-                                let is_expand_all = indices.len() == 1
-                                    && matches!(indices[0].kind, HirExprKind::Colon);
-                                if is_expand_all {
-                                    specs.push(crate::instr::ArgSpec {
-                                        is_expand: true,
-                                        num_indices: 0,
-                                        expand_all: true,
-                                    });
-                                    self.compile_expr(base)?;
-                                } else {
-                                    specs.push(crate::instr::ArgSpec {
-                                        is_expand: true,
-                                        num_indices: indices.len(),
-                                        expand_all: false,
-                                    });
-                                    self.compile_expr(base)?;
-                                    for i in indices {
-                                        self.compile_expr(i)?;
-                                    }
-                                }
-                            } else {
-                                specs.push(crate::instr::ArgSpec {
-                                    is_expand: false,
-                                    num_indices: 0,
-                                    expand_all: false,
-                                });
-                                self.compile_expr(arg)?;
-                            }
-                        }
-                        self.emit_call_with_arg_spans(
-                            Instr::CallFunctionExpandMulti(name.clone(), specs),
-                            &call_arg_spans,
-                        );
-                    } else {
-                        for arg in args {
-                            self.compile_expr(arg)?;
-                        }
-                        self.emit_call_with_arg_spans(
-                            Instr::CallFunction(name.clone(), args.len()),
-                            &call_arg_spans,
-                        );
-                    }
-                } else {
-                    // Existing import-based function/builtin resolution, extended with static method via Class.*
-                    // Attempt compile-time import resolution for builtins and user functions with ambiguity checks
-                    // Precedence for unqualified resolution:
-                    // locals > user functions in scope > specific imports > wildcard imports > Class.* static methods
-                    // 1) Specific imports: import pkg.foo => resolve 'foo' (takes precedence over wildcard)
-                    // 2) Wildcard imports: import pkg.* => resolve 'pkg.foo'
-                    // 3) Class.* static methods: import Class.* (or pkg.Class.*) => resolve static methods if unambiguous
-                    let mut resolved = name.clone();
-                    let mut static_candidates: Vec<(String, String)> = Vec::new();
-                    if !runmat_builtins::builtin_functions()
-                        .iter()
-                        .any(|b| b.name == resolved)
-                    {
-                        // Specific candidates
-                        let mut specific_candidates: Vec<String> = Vec::new();
-                        for (path, wildcard) in &self.imports {
-                            if *wildcard {
-                                continue;
-                            }
-                            if path.last().map(|s| s.as_str()) == Some(name.as_str()) {
-                                let qual = path.join(".");
-                                if runmat_builtins::builtin_functions()
-                                    .iter()
-                                    .any(|b| b.name == qual)
-                                    || self.functions.contains_key(&qual)
-                                {
-                                    specific_candidates.push(qual);
-                                }
-                            }
-                        }
-                        if specific_candidates.len() > 1 {
-                            return Err(self.compile_error(format!(
-                                "ambiguous unqualified reference '{}' via imports: {}",
-                                name,
-                                specific_candidates.join(", ")
-                            )));
-                        }
-                        if specific_candidates.len() == 1 {
-                            resolved = specific_candidates.remove(0);
-                        } else {
-                            // Wildcard candidates for functions
-                            let mut wildcard_candidates: Vec<String> = Vec::new();
-                            for (path, wildcard) in &self.imports {
-                                if !*wildcard {
-                                    continue;
-                                }
-                                if path.is_empty() {
-                                    continue;
-                                }
-                                let mut qual = String::new();
-                                for (i, part) in path.iter().enumerate() {
-                                    if i > 0 {
-                                        qual.push('.');
-                                    }
-                                    qual.push_str(part);
-                                }
-                                qual.push('.');
-                                qual.push_str(name);
-                                if runmat_builtins::builtin_functions()
-                                    .iter()
-                                    .any(|b| b.name == qual)
-                                    || self.functions.contains_key(&qual)
-                                {
-                                    wildcard_candidates.push(qual);
-                                }
-                                // Accumulate Class.* static method candidates for any class path
-                                let mut cls = String::new();
-                                for (i, part) in path.iter().enumerate() {
-                                    if i > 0 {
-                                        cls.push('.');
-                                    }
-                                    cls.push_str(part);
-                                }
-                                if let Some((m, _owner)) =
-                                    runmat_builtins::lookup_method(&cls, name)
-                                {
-                                    if m.is_static {
-                                        static_candidates.push((cls.clone(), name.clone()));
-                                    }
-                                }
-                            }
-                            if wildcard_candidates.len() > 1 {
-                                return Err(self.compile_error(format!(
-                                    "ambiguous unqualified reference '{}' via wildcard imports: {}",
-                                    name,
-                                    wildcard_candidates.join(", ")
-                                )));
-                            }
-                            if wildcard_candidates.len() == 1 {
-                                resolved = wildcard_candidates.remove(0);
-                            }
-                        }
-                    }
-                    // If resolved maps to a user function, compile it now
-                    if self.functions.contains_key(&resolved) {
-                        if has_any_expand {
-                            let mut specs: Vec<crate::instr::ArgSpec> =
-                                Vec::with_capacity(args.len());
-                            for arg in args {
-                                if let HirExprKind::IndexCell(base, indices) = &arg.kind {
-                                    let is_expand_all = indices.len() == 1
-                                        && matches!(indices[0].kind, HirExprKind::Colon);
-                                    if is_expand_all {
-                                        specs.push(crate::instr::ArgSpec {
-                                            is_expand: true,
-                                            num_indices: 0,
-                                            expand_all: true,
-                                        });
-                                        self.compile_expr(base)?;
-                                    } else {
-                                        specs.push(crate::instr::ArgSpec {
-                                            is_expand: true,
-                                            num_indices: indices.len(),
-                                            expand_all: false,
-                                        });
-                                        self.compile_expr(base)?;
-                                        for i in indices {
-                                            self.compile_expr(i)?;
-                                        }
-                                    }
-                                } else {
-                                    specs.push(crate::instr::ArgSpec {
-                                        is_expand: false,
-                                        num_indices: 0,
-                                        expand_all: false,
-                                    });
-                                    self.compile_expr(arg)?;
-                                }
-                            }
-                            self.emit_call_with_arg_spans(
-                                Instr::CallFunctionExpandMulti(resolved.clone(), specs),
-                                &call_arg_spans,
-                            );
-                            return Ok(());
-                        } else {
-                            // Flatten inner user-function returns into argument list for the call
-                            let mut total_argc: usize = 0;
-                            for arg in args {
-                                if let HirExprKind::FuncCall(inner, inner_args) = &arg.kind {
-                                    if self.functions.contains_key(inner) {
-                                        for a in inner_args {
-                                            self.compile_expr(a)?;
-                                        }
-                                        let outc = self
-                                            .functions
-                                            .get(inner)
-                                            .map(|f| f.outputs.len().max(1))
-                                            .unwrap_or(1);
-                                        self.emit(Instr::CallFunctionMulti(
-                                            inner.clone(),
-                                            inner_args.len(),
-                                            outc,
-                                        ));
-                                        self.emit(Instr::Unpack(outc));
-                                        total_argc += outc;
-                                        continue;
-                                    }
-                                }
-                                self.compile_expr(arg)?;
-                                total_argc += 1;
-                            }
-                            self.emit_call_with_arg_spans(
-                                Instr::CallFunction(resolved.clone(), total_argc),
-                                &call_arg_spans,
-                            );
-                            return Ok(());
-                        }
-                    }
-                    // If still no function, and exactly one static candidate, call it
-                    if !runmat_builtins::builtin_functions()
-                        .iter()
-                        .any(|b| b.name == resolved)
-                        && static_candidates.len() == 1
-                    {
-                        let (cls, method) = static_candidates.remove(0);
-                        for arg in args {
-                            self.compile_expr(arg)?;
-                        }
-                        self.emit_call_with_arg_spans(
-                            Instr::CallStaticMethod(cls, method, args.len()),
-                            &call_arg_spans,
-                        );
-                        return Ok(());
-                    }
-                    // If multiple static candidates and no function resolved, report ambiguity
-                    if !runmat_builtins::builtin_functions()
-                        .iter()
-                        .any(|b| b.name == resolved)
-                        && static_candidates.len() > 1
-                    {
-                        return Err(self.compile_error(format!(
-                            "ambiguous unqualified static method '{}' via Class.* imports: {}",
-                            name,
-                            static_candidates
-                                .iter()
-                                .map(|(c, _)| c.clone())
-                                .collect::<Vec<_>>()
-                                .join(", ")
-                        )));
-                    }
-                    // Existing propagation path and builtin call
-                    if !has_any_expand {
-                        // First scan for user-defined inner function calls to expand; avoid compiling
-                        // simple args here to prevent duplicating them on the stack. If any inner
-                        // user functions are found, compile them and then compile the remaining
-                        // simple args once, emit the builtin call, and return. Otherwise, fall
-                        // through to the normal (single-pass) compilation below.
-                        let mut total_argc = 0usize;
-                        let mut did_expand_inner = false;
-                        let mut pending_simple: Vec<&runmat_hir::HirExpr> = Vec::new();
-                        for arg in args {
-                            if let HirExprKind::FuncCall(inner, inner_args) = &arg.kind {
-                                if self.functions.contains_key(inner) {
-                                    for a in inner_args {
-                                        self.compile_expr(a)?;
-                                    }
-                                    let outc = self
-                                        .functions
-                                        .get(inner)
-                                        .map(|f| f.outputs.len().max(1))
-                                        .unwrap_or(1);
-                                    self.emit(Instr::CallFunctionMulti(
-                                        inner.clone(),
-                                        inner_args.len(),
-                                        outc,
-                                    ));
-                                    self.emit(Instr::Unpack(outc));
-                                    total_argc += outc;
-                                    did_expand_inner = true;
-                                } else {
-                                    pending_simple.push(arg);
-                                }
-                            } else {
-                                pending_simple.push(arg);
-                            }
-                        }
-                        if did_expand_inner {
-                            for arg in pending_simple {
-                                self.compile_expr(arg)?;
-                                total_argc += 1;
-                            }
-                            self.emit_call_with_arg_spans(
-                                Instr::CallBuiltin(resolved, total_argc),
-                                &call_arg_spans,
-                            );
-                            return Ok(());
-                        }
-                    }
-                    if has_any_expand {
-                        let mut specs: Vec<crate::instr::ArgSpec> = Vec::with_capacity(args.len());
-                        for arg in args {
-                            if let HirExprKind::IndexCell(base, indices) = &arg.kind {
-                                let is_expand_all = indices.len() == 1
-                                    && matches!(indices[0].kind, HirExprKind::Colon);
-                                if is_expand_all {
-                                    specs.push(crate::instr::ArgSpec {
-                                        is_expand: true,
-                                        num_indices: 0,
-                                        expand_all: true,
-                                    });
-                                    self.compile_expr(base)?;
-                                } else {
-                                    specs.push(crate::instr::ArgSpec {
-                                        is_expand: true,
-                                        num_indices: indices.len(),
-                                        expand_all: false,
-                                    });
-                                    self.compile_expr(base)?;
-                                    for i in indices {
-                                        self.compile_expr(i)?;
-                                    }
-                                }
-                            } else {
-                                specs.push(crate::instr::ArgSpec {
-                                    is_expand: false,
-                                    num_indices: 0,
-                                    expand_all: false,
-                                });
-                                self.compile_expr(arg)?;
-                            }
-                        }
-                        self.emit_call_with_arg_spans(
-                            Instr::CallBuiltinExpandMulti(resolved, specs),
-                            &call_arg_spans,
-                        );
-                    } else {
-                        for arg in args {
-                            self.compile_expr(arg)?;
-                        }
-                        self.emit_call_with_arg_spans(
-                            Instr::CallBuiltin(resolved, args.len()),
-                            &call_arg_spans,
-                        );
-                    }
-                    return Ok(());
-                }
-                return Ok(());
-            }
-            HirExprKind::Tensor(matrix_data) | HirExprKind::Cell(matrix_data) => {
-                let rows = matrix_data.len();
-                // Special case: 1-row tensor literal with a single element that is IndexCell(base, {:})
-                // Lower "[C{:}]" into cat(2, C{:}) so downstream expansion works without colon compilation
-                if matches!(expr.kind, HirExprKind::Tensor(_))
-                    && rows == 1
-                    && matrix_data.first().map(|r| r.len()).unwrap_or(0) == 1
-                {
-                    if let HirExprKind::IndexCell(base, indices) = &matrix_data[0][0].kind {
-                        if indices.len() == 1 && matches!(indices[0].kind, HirExprKind::Colon) {
-                            // Build specs: first fixed dim=2, then expand_all for base
-                            let mut specs: Vec<crate::instr::ArgSpec> = Vec::with_capacity(2);
-                            // Fixed dimension 2
-                            specs.push(crate::instr::ArgSpec {
-                                is_expand: false,
-                                num_indices: 0,
-                                expand_all: false,
-                            });
-                            self.emit(Instr::LoadConst(2.0));
-                            // Expand all from base cell
-                            specs.push(crate::instr::ArgSpec {
-                                is_expand: true,
-                                num_indices: 0,
-                                expand_all: true,
-                            });
-                            self.compile_expr(base)?;
-                            self.emit(Instr::CallBuiltinExpandMulti("cat".to_string(), specs));
-                            return Ok(());
-                        }
-                    }
-                }
-                let has_non_literals = matrix_data.iter().any(|row| {
-                    row.iter()
-                        .any(|expr| !matches!(expr.kind, HirExprKind::Number(_)))
-                });
-                if has_non_literals {
-                    for row in matrix_data {
-                        for element in row {
-                            self.compile_expr(element)?;
-                        }
-                    }
-                    let row_lengths: Vec<usize> = matrix_data.iter().map(|row| row.len()).collect();
-                    if matches!(expr.kind, HirExprKind::Cell(_)) {
-                        // For 2D cells, we know rows and row lengths; emit 2D version when rectangular
-                        let rectangular = row_lengths.iter().all(|&c| c == row_lengths[0]);
-                        if rectangular {
-                            let cols = if rows > 0 { row_lengths[0] } else { 0 };
-                            self.emit(Instr::CreateCell2D(rows, cols));
-                        } else {
-                            // Ragged cells: fall back to 1D create with total count and row-major ordering
-                            let total: usize = row_lengths.iter().sum();
-                            self.emit(Instr::CreateCell2D(1, total));
-                        }
-                    } else {
-                        for &row_len in &row_lengths {
-                            self.emit(Instr::LoadConst(row_len as f64));
-                        }
-                        self.emit(Instr::CreateMatrixDynamic(rows));
-                    }
-                } else {
-                    let cols = if rows > 0 { matrix_data[0].len() } else { 0 };
-                    for row in matrix_data {
-                        for element in row {
-                            self.compile_expr(element)?;
-                        }
-                    }
-                    if matches!(expr.kind, HirExprKind::Cell(_)) {
-                        self.emit(Instr::CreateCell2D(rows, cols));
-                    } else {
-                        self.emit(Instr::CreateMatrix(rows, cols));
-                    }
-                }
-            }
-            HirExprKind::Index(base, indices) => {
-                let has_colon = indices.iter().any(|e| matches!(e.kind, HirExprKind::Colon));
-                let has_end = indices.iter().any(Self::expr_contains_end);
-                let has_vector = indices.iter().any(|e| {
-                    matches!(e.kind, HirExprKind::Range(_, _, _) | HirExprKind::Tensor(_))
-                        || matches!(
-                            e.ty,
-                            runmat_hir::Type::Tensor { .. }
-                                | runmat_hir::Type::Bool
-                                | runmat_hir::Type::Logical { .. }
-                        )
-                        || !matches!(
-                            e.kind,
-                            HirExprKind::Number(_) | HirExprKind::Colon | HirExprKind::End
-                        )
-                });
-                // General case: any-dimension ranges with end arithmetic (e.g., A(:,2:2:end-1,...))
-                // We lower into IndexSliceExpr: push base, then per-range start[, step] in increasing dimension order,
-                // then any numeric scalar indices (in order). Colon and plain end dims are marked in masks.
-                {
-                    let mut has_any_range_end = false;
-                    let mut range_dims: Vec<usize> = Vec::new();
-                    let mut range_has_step: Vec<bool> = Vec::new();
-                    let mut range_start_exprs: Vec<Option<EndExpr>> = Vec::new();
-                    let mut range_step_exprs: Vec<Option<EndExpr>> = Vec::new();
-                    let mut end_offsets: Vec<EndExpr> = Vec::new();
-                    // First pass: detect any Range whose upper bound depends on `end`.
-                    for (dim, index) in indices.iter().enumerate() {
-                        if let HirExprKind::Range(_start, step, _end) = &index.kind {
-                            if let Some((start_end, step_end, end_expr)) =
-                                range_dynamic_end_spec(index)
-                            {
-                                has_any_range_end = true;
-                                range_dims.push(dim);
-                                range_has_step.push(step.is_some());
-                                range_start_exprs.push(start_end);
-                                range_step_exprs.push(step_end);
-                                end_offsets.push(end_expr);
-                            }
-                        }
-                    }
-                    if has_any_range_end {
-                        self.compile_expr(base)?;
-                        // Push per-range start and optional step in dimension order
-                        for (ri, &dim) in range_dims.iter().enumerate() {
-                            if let HirExprKind::Range(start, step, _end) = &indices[dim].kind {
-                                if range_start_exprs[ri].is_some() {
-                                    self.emit(Instr::LoadConst(0.0));
-                                } else {
-                                    self.compile_expr(start)?;
-                                }
-                                if let Some(st) = step {
-                                    if range_step_exprs[ri].is_some() {
-                                        self.emit(Instr::LoadConst(0.0));
-                                    } else {
-                                        self.compile_expr(st)?;
-                                    }
-                                }
-                            }
-                        }
-                        // Count numeric scalar indices and push them
-                        let mut colon_mask: u32 = 0;
-                        let mut end_mask: u32 = 0;
-                        let mut numeric_count = 0usize;
-                        for (dim, index) in indices.iter().enumerate() {
-                            match &index.kind {
-                                HirExprKind::Colon => {
-                                    colon_mask |= 1u32 << dim;
-                                }
-                                HirExprKind::End => {
-                                    end_mask |= 1u32 << dim;
-                                }
-                                HirExprKind::Range(_, _, _) => {
-                                    // End-bounded ranges are encoded via `range_dims`; all other
-                                    // ranges still need their compiled selector payload.
-                                    if range_dynamic_end_spec(index).is_some() {
-                                        continue;
-                                    }
-                                    self.compile_expr(index)?;
-                                    numeric_count += 1;
-                                }
-                                _ => {
-                                    self.compile_expr(index)?;
-                                    numeric_count += 1;
-                                }
-                            }
-                        }
-                        self.emit(Instr::IndexSliceExpr {
-                            dims: indices.len(),
-                            numeric_count,
-                            colon_mask,
-                            end_mask,
-                            range_dims,
-                            range_has_step,
-                            range_start_exprs,
-                            range_step_exprs,
-                            range_end_exprs: end_offsets,
-                            end_numeric_exprs: Vec::new(),
-                        });
-                        return Ok(());
-                    }
-                }
-                if has_colon
-                    || has_vector
-                    || has_end
-                    || indices.len() > 2
-                    || Self::expr_contains_end(base)
-                {
-                    // Push base first, then numeric indices in order; compute colon mask
-                    self.compile_expr(base)?;
-                    let mut colon_mask: u32 = 0;
-                    let mut end_mask: u32 = 0;
-                    let mut numeric_count = 0usize;
-                    let mut end_offsets: Vec<(usize, EndExpr)> = Vec::new();
-                    for (dim, index) in indices.iter().enumerate() {
-                        if matches!(index.kind, HirExprKind::Colon) {
-                            colon_mask |= 1u32 << dim;
-                        } else if matches!(index.kind, HirExprKind::End) {
-                            end_mask |= 1u32 << dim;
-                        } else {
-                            if let Some(off) = end_numeric_expr(index) {
-                                // Reserve a numeric slot: push placeholder and count it.
-                                // Positive value -> end-k, negative value -> floor(end/|k|).
-                                self.emit(Instr::LoadConst(0.0));
-                                end_offsets.push((numeric_count, off));
-                                numeric_count += 1;
-                                continue;
-                            }
-                            self.compile_expr(index)?;
-                            numeric_count += 1;
-                        }
-                    }
-                    if end_offsets.is_empty() {
-                        self.emit(Instr::IndexSlice(
-                            indices.len(),
-                            numeric_count,
-                            colon_mask,
-                            end_mask,
-                        ));
-                    } else {
-                        self.emit(Instr::IndexSliceExpr {
-                            dims: indices.len(),
-                            numeric_count,
-                            colon_mask,
-                            end_mask,
-                            range_dims: Vec::new(),
-                            range_has_step: Vec::new(),
-                            range_start_exprs: Vec::new(),
-                            range_step_exprs: Vec::new(),
-                            range_end_exprs: Vec::new(),
-                            end_numeric_exprs: end_offsets,
-                        });
-                    }
-                } else {
-                    self.compile_expr(base)?;
-                    for index in indices {
-                        self.compile_expr(index)?;
-                    }
-                    self.emit(Instr::Index(indices.len()));
-                }
-            }
-            HirExprKind::Colon => {
-                // Placeholder for contexts where colon appeared in RHS expansion; real colon handling occurs in indexing logic
-                // Emit a benign constant to keep stack discipline when mistakenly compiled
-                self.emit(Instr::LoadConst(0.0));
-            }
-            HirExprKind::End => {
-                self.emit(Instr::LoadConst(-0.0)); /* placeholder, resolved via end_mask in IndexSlice */
-            }
-            HirExprKind::Member(base, field) => {
-                // If base is a known class ref literal (string via classref builtin), static access
-                // Or if base is MetaClass (string literal), treat as class name for static access
-                // Otherwise, instance member
-                match &base.kind {
-                    HirExprKind::MetaClass(cls_name) => {
-                        self.emit(Instr::LoadStaticProperty(cls_name.clone(), field.clone()));
-                    }
-                    HirExprKind::FuncCall(name, args) if name == "classref" && args.len() == 1 => {
-                        self.compile_expr(base)?;
-                        self.emit(Instr::LoadMember(field.clone()));
-                    }
-                    _ => {
-                        // Default to instance property access; subsref overloading is handled at runtime via call_method if needed
-                        self.compile_expr(base)?;
-                        self.emit(Instr::LoadMember(field.clone()));
-                    }
-                }
-            }
-            HirExprKind::MemberDynamic(base, name_expr) => {
-                self.compile_expr(base)?;
-                self.compile_expr(name_expr)?;
-                self.emit(Instr::LoadMemberDynamic);
-            }
-            HirExprKind::DottedInvoke(base, member, args) => {
-                self.compile_expr(base)?;
-                for arg in args {
-                    self.compile_expr(arg)?;
-                }
-                self.emit(Instr::CallMethodOrMemberIndex(member.clone(), args.len()));
-            }
-            // Dynamic member s.(expr)
-            HirExprKind::MethodCall(b, m, a) if m == &"()".to_string() && a.len() == 1 => {
-                // Note: parser currently doesn't produce this form; placeholder for dynamic
-                self.compile_expr(b)?;
-                self.compile_expr(&a[0])?;
-                self.emit(Instr::LoadMemberDynamic);
-            }
-            HirExprKind::MethodCall(base, method, args) => match &base.kind {
-                HirExprKind::MetaClass(cls_name) => {
-                    for arg in args {
-                        self.compile_expr(arg)?;
-                    }
-                    self.emit(Instr::CallStaticMethod(
-                        cls_name.clone(),
-                        method.clone(),
-                        args.len(),
-                    ));
-                }
-                HirExprKind::FuncCall(name, bargs) if name == "classref" && bargs.len() == 1 => {
-                    if let HirExprKind::String(cls) = &bargs[0].kind {
-                        let cls_name = Self::normalize_class_literal_name(cls);
-                        for arg in args {
-                            self.compile_expr(arg)?;
-                        }
-                        self.emit(Instr::CallStaticMethod(
-                            cls_name,
-                            method.clone(),
-                            args.len(),
-                        ));
-                    } else {
-                        self.compile_expr(base)?;
-                        for arg in args {
-                            self.compile_expr(arg)?;
-                        }
-                        self.emit(Instr::CallMethod(method.clone(), args.len()));
-                    }
-                }
-                _ => {
-                    self.compile_expr(base)?;
-                    for arg in args {
-                        self.compile_expr(arg)?;
-                    }
-                    self.emit(Instr::CallMethod(method.clone(), args.len()));
-                }
-            },
-            HirExprKind::AnonFunc { params, body } => self.compile_anon_func(params, body)?,
-            HirExprKind::FuncHandle(name) => {
-                self.emit(Instr::LoadString(name.clone()));
-                self.emit(Instr::CallBuiltin("make_handle".to_string(), 1));
-            }
-            HirExprKind::MetaClass(name) => {
-                self.emit(Instr::LoadString(name.clone()));
-            }
-            // Member/Method on metaclass (string on stack) will be handled by runtime as static property/method via classref
-            // We lower MetaClass to a string (class name) and then member/method code paths remain unchanged.
-            HirExprKind::IndexCell(base, indices) => {
-                self.compile_expr(base)?;
-                for index in indices {
-                    self.compile_expr(index)?;
-                }
-                self.emit(Instr::IndexCell(indices.len()));
-            }
-        }
-        Ok(())
-    }
-}
-
-fn label_for_expr(expr: &HirExpr) -> EmitLabel {
-    if let HirExprKind::Var(var_id) = &expr.kind {
-        EmitLabel::Var(var_id.0)
-    } else {
-        EmitLabel::Ans
+        self.compile_expr_impl(expr)
     }
 }
